third_party/WebKit/LayoutTests/animations/animationworklet/peek-updated-composited-property-on-main-expected.txt

CONSOLE MESSAGE: line 44: opacity of target is 0.5
